:label: (core): 将 minVersion 类型从 Int 改为 Long
修改 AutoHooker 类中的 minVersion 属性类型，从 Int 变更为 Long 以支持更大的版本号范围，提高版本控制的灵活性和兼容性。

+++ app/src/main/java/net/ankio/auto/xposed/hooks/auto/AutoHooker.kt
@@ -26,7 +26,7 @@ import net.ankio.dex.model.Clazz
 
 
 class AutoHooker : HookerManifest() {
-    override var minVersion: Int = 212
+    override var minVersion: Long = 212
     override val packageName: String
         get() = BuildConfig.APPLICATION_ID
     override val appName: String = "自动记账"
